SplunkSearch.py
#Read data from a file named “DataList.txt” which contains a horizontal
list of data (i.e. IPs, Domains, etc.) and save them in Splunk friendly
format in a new file called “FormatedDataList.txt”. The format of the
new file will be: “data” OR “data” OR “data” OR ….
The logical operator of OR could easily be replaced with another
operator (i.e. AND) by making the switch in the program.